feat: update local history detail
