+++ apps/mobile/src/screens/Transaction/components/Actions/RevokeNFT.tsx
@@ -1,7 +1,5 @@
 /* eslint-disable react-native/no-inline-styles */
-import { RcIconExternalLinkCC, RcIconRightCC } from '@/assets/icons/common';
 import RcIconSingleArrow from '@/assets2024/icons/history/IconSingleArrow.svg';
-import ChainIconImage from '@/components/Chain/ChainIconImage';
 import { useTheme2024 } from '@/hooks/theme';
 import { findChain } from '@/utils/chain';
 import { createGetStyles2024 } from '@/utils/styles';
@@ -11,28 +9,24 @@ import { TouchableOpacity, View } from 'react-native';
 
 import { TransactionGroup } from '@/core/services/transactionHistory';
 
-import ViewMore from '@/components/Approval/components/Actions/components/ViewMore';
-import { AssetAvatar } from '@/components/AssetAvatar';
-import { toast } from '@/components2024/Toast';
 import { RootNames } from '@/constant/layout';
 import { useAccounts } from '@/hooks/account';
 import { useSortAddressList } from '@/screens/Address/useSortAddressList';
-import { TransactionPendingDetail } from '@/screens/TransactionRecord/components/TransactionPendingDetail';
-import { ellipsisAddress } from '@/utils/address';
 import { naviPush } from '@/utils/navigation';
-import { formatAmount } from '@/utils/number';
-import { formatIntlTimestamp } from '@/utils/time';
-import { openTxExternalUrl } from '@/utils/transaction';
 import { RevokeNFTRequireData } from '@rabby-wallet/rabby-action';
 import { useMemoizedFn } from 'ahooks';
 import { unionBy } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { AddressItemInDetail } from '../../HistoryDetailScreen';
-import { TxStatusItem } from '../TxStatusItem';
 import { HistoryItemIcon } from '../HistoryItemIcon';
 import { HistoryItemCateType } from '../type';
 import { Account } from '@/core/services/preference';
 import { Text } from '@/components/Typography';
+import {
+  ActionDetailItem,
+  ActionDetailSection,
+  ActionDetailText,
+} from './components/ActionDetailSection';
+import { ActionSpenderView } from './components/ActionSpenderView';
 
 interface Props {
   data: TransactionGroup;
@@ -72,16 +66,6 @@ export const RevokeNFT: React.FC<Props> = ({
   const unionAccounts = useMemo(() => {
     return unionBy(list, account => account.address.toLowerCase());
   }, [list]);
-
-  const handleOpenTxId = useMemoizedFn(() => {
-    const tx = data.maxGasTx.hash;
-
-    if (chain?.scanLink) {
-      openTxExternalUrl({ chain, txHash: tx });
-    } else {
-      toast.error('Unknown chain');
-    }
-  });
 
   const handleGotoDetail = useMemoizedFn(() => {
     naviPush(RootNames.NftDetail, {
@@ -132,265 +116,39 @@ export const RevokeNFT: React.FC<Props> = ({
             <Text style={styles.itemTitleText}>
               {t('page.transactions.detail.RevokeFrom')}
             </Text>
-            <ViewMore
-              type="nftSpender"
-              data={{
-                ...requireData,
-                spender: actionData.spender,
-                chain,
-              }}>
-              <View style={{ alignItems: 'flex-end' }}>
-                <View
-                  style={{
-                    flexDirection: 'row',
-                    alignItems: 'center',
-                    gap: 4,
-                  }}>
-                  <AssetAvatar
-                    logo={requireData?.protocol?.logo_url}
-                    size={16}
-                  />
-                  <Text style={[styles.itemContentText]}>
-                    {requireData?.protocol?.name || t('global.Unknown')}
-                  </Text>
-                  <RcIconRightCC
-                    width={14}
-                    height={14}
-                    color={colors2024['neutral-foot']}
-                  />
-                </View>
-                <Text style={styles.itemAddressText}>
-                  {ellipsisAddress(actionData.spender)}
-                </Text>
-              </View>
-            </ViewMore>
+            <ActionSpenderView
+              requireData={requireData}
+              spender={actionData.spender}
+              chain={chain}
+            />
           </View>
         </View>
       </TouchableOpacity>
-      <View style={styles.detailContainer}>
-        <View style={styles.detailContainerHeader}>
-          <Text style={styles.detailContainerTitle}>
-            {t('page.transactions.detail.TransactionDetails')}
-          </Text>
-        </View>
-        {!data.isPending && data.maxGasTx.completedAt && (
-          <View style={styles.detailItem}>
-            <Text style={styles.itemTitleText}>
-              {t('page.transactions.detail.Date')}
-            </Text>
-            <View>
-              <Text style={styles.itemContentText}>
-                {formatIntlTimestamp(data?.maxGasTx.completedAt)}
-              </Text>
-            </View>
-          </View>
-        )}
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.Status')}
-          </Text>
-          <View>
-            <TxStatusItem
-              status={data.isFailed ? 0 : 1}
-              isPending={data.isPending}
-              withText={true}
-            />
-          </View>
-        </View>
-        {data.isPending ? <TransactionPendingDetail data={data} /> : null}
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.RevokeFrom')}
-          </Text>
-          <ViewMore
-            type="nftSpender"
-            data={{
-              ...requireData,
-              spender: actionData.spender,
-              chain,
-            }}>
-            <View style={{ alignItems: 'flex-end' }}>
-              <View
-                style={{
-                  flexDirection: 'row',
-                  alignItems: 'center',
-                  gap: 4,
-                }}>
-                <AssetAvatar logo={requireData?.protocol?.logo_url} size={16} />
-                <Text style={[styles.itemContentText]}>
-                  {requireData?.protocol?.name || t('global.Unknown')}
-                </Text>
-                <RcIconRightCC
-                  width={14}
-                  height={14}
-                  color={colors2024['neutral-foot']}
-                />
-              </View>
-              <Text style={styles.itemAddressText}>
-                {ellipsisAddress(actionData.spender)}
-              </Text>
-            </View>
-          </ViewMore>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.name')}
-          </Text>
-          <Text
-            style={styles.itemContentText}
-            numberOfLines={1}
-            ellipsizeMode="tail">
-            {actionData?.nft?.name || '-'}
-          </Text>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.Chain')}
-          </Text>
-          <View style={{ flexDirection: 'row', gap: 4 }}>
-            <ChainIconImage
-              size={16}
-              chainEnum={chain?.enum}
-              isShowRPCStatus={true}
-            />
-            <Text style={[styles.itemContentText]}>{chain?.name}</Text>
-          </View>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.From')}
-          </Text>
-          <AddressItemInDetail
-            address={data.maxGasTx.address}
-            accounts={unionAccounts}
+      <ActionDetailSection data={data} chain={chain} accounts={unionAccounts}>
+        <ActionDetailItem label={t('page.transactions.detail.RevokeFrom')}>
+          <ActionSpenderView
+            requireData={requireData}
+            spender={actionData.spender}
+            chain={chain}
           />
-        </View>
+        </ActionDetailItem>
 
-        {Boolean(data.maxGasTx?.gasUSDValue) && (
-          <View style={styles.detailItem}>
-            <Text style={styles.itemTitleText}>
-              {t('page.transactions.detail.GasFee')}
-            </Text>
-            <Text style={styles.itemContentText}>
-              {formatAmount(data.maxGasTx?.gasTokenCount!)}{' '}
-              {data.maxGasTx?.gasTokenSymbol || ''} ($
-              {formatAmount(data.maxGasTx?.gasUSDValue ?? 0)})
-            </Text>
-          </View>
-        )}
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>Hash</Text>
-          <TouchableOpacity
-            disabled={!chain?.scanLink}
-            onPress={handleOpenTxId}
-            style={{ flexDirection: 'row', alignItems: 'center', gap: 2 }}>
-            <Text style={[styles.itemContentText]}>
-              {ellipsisAddress(data.maxGasTx.hash!)}
-            </Text>
-            <RcIconExternalLinkCC
-              width={14}
-              height={14}
-              color={colors2024['neutral-foot']}
-            />
-          </TouchableOpacity>
-        </View>
-      </View>
+        <ActionDetailItem label={t('page.transactions.detail.name')}>
+          <ActionDetailText numberOfLines={1} ellipsizeMode="tail">
+            {actionData?.nft?.name || '-'}
+          </ActionDetailText>
+        </ActionDetailItem>
+      </ActionDetailSection>
     </>
   );
 };
 
 const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
-  detailContainer: {
-    // flex: 1,
-    width: '100%',
-    marginTop: 12,
-    borderRadius: 16,
-    paddingTop: 12,
-    paddingBottom: 4,
-    backgroundColor: !isLight
-      ? colors2024['neutral-bg-2']
-      : colors2024['neutral-bg-1'],
-  },
-  detailContainerHeader: {
-    marginBottom: 8,
-    paddingHorizontal: 16,
-  },
-  detailContainerTitle: {
-    color: colors2024['neutral-body'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-  },
-  ghostButton: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderColor: colors2024['neutral-info'],
-  },
-  primaryButton: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderColor: colors2024['brand-default'],
-  },
-  primaryTitle: {
-    color: colors2024['brand-default'],
-  },
-  ghostTitle: {
-    color: colors2024['neutral-title-1'],
-  },
-  iconSwitchArrow: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderRadius: 200,
-    width: 45,
-    height: 45,
-    alignItems: 'center',
-    justifyContent: 'center',
-    position: 'absolute',
-    left: '50%',
-    top: '50%',
-    marginLeft: -22,
-    marginTop: -22,
-  },
-  tokenAmountTextList: {
-    color: colors2024['green-default'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 18,
-    lineHeight: 22,
-    fontWeight: '700',
-  },
   colomnBox: {
     flexDirection: 'column',
   },
   isSendTextColor: {
     color: colors2024['neutral-title-1'],
-  },
-  isFailBox: {
-    opacity: 0.3,
-  },
-  image: {
-    width: 46,
-    height: 46,
-  },
-  fromTokenBox: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 16,
-    backgroundColor: colors2024['neutral-bg-1'],
-    flex: 1,
-    height: 110,
-    gap: 10,
-  },
-  toTokenBox: {
-    gap: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 16,
-    backgroundColor: colors2024['neutral-bg-1'],
-    flex: 1,
-    height: 110,
   },
   card: {
     width: '100%',
@@ -420,40 +178,7 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     fontWeight: '500',
     marginTop: 2,
   },
-  mutliBox: {
-    width: '100%',
-    backgroundColor: colors2024['neutral-bg-1'],
-    justifyContent: 'center',
-    alignContent: 'center',
-    borderRadius: 16,
-    paddingHorizontal: 12,
-    paddingVertical: 16,
-    // flexDirection: 'row',
-    gap: 12,
-  },
-  doubleBox: {
-    justifyContent: 'center',
-    alignContent: 'center',
-    flexDirection: 'row',
-    height: 110,
-    gap: 10,
-    position: 'relative',
-  },
 
-  buttonContainer: {
-    position: 'absolute',
-    flexDirection: 'row',
-    height: 60,
-    bottom: 40,
-    width: '100%',
-    gap: 16,
-    left: 16,
-  },
-  itemAliaName: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 4,
-  },
   extraItem: {
     flexDirection: 'row',
     padding: 12,
@@ -464,14 +189,6 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     marginHorizontal: 12,
     marginBottom: 12,
   },
-  detailItem: {
-    flexDirection: 'row',
-    gap: 8,
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
   itemTitleText: {
     color: colors2024['neutral-secondary'],
     fontFamily: 'SF Pro Rounded',
@@ -480,36 +197,4 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     fontWeight: '500',
     maxWidth: '45%',
   },
-  itemAddressText: {
-    color: colors2024['neutral-secondary'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '400',
-  },
-  itemContentText: {
-    color: colors2024['neutral-body'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-  },
-  headerTitleStyle: {
-    color: colors2024['neutral-title-1'],
-    fontWeight: '800',
-    fontSize: 20,
-    fontFamily: 'SF Pro Rounded',
-    lineHeight: 24,
-  },
-
-  statuItemText: {
-    color: colors2024['green-default'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-    marginLeft: 4,
-  },
-
-  headerItem: {},
 }));
