+++ apps/mobile/src/screens/Transaction/components/Actions/RevokeToken.tsx
@@ -1,10 +1,7 @@
 /* eslint-disable react-native/no-inline-styles */
-import { RcIconExternalLinkCC, RcIconRightCC } from '@/assets/icons/common';
 import RcIconSingleArrow from '@/assets2024/icons/history/IconSingleArrow.svg';
-import ChainIconImage from '@/components/Chain/ChainIconImage';
 import { useTheme2024 } from '@/hooks/theme';
 import { findChain } from '@/utils/chain';
-import { formatAmount } from '@/utils/number';
 import { createGetStyles2024 } from '@/utils/styles';
 import { getTokenSymbol, tokenItemToITokenItem } from '@/utils/token';
 import { TokenItem } from '@rabby-wallet/rabby-api/dist/types';
@@ -13,27 +10,24 @@ import { TouchableOpacity, View } from 'react-native';
 
 import { TransactionGroup } from '@/core/services/transactionHistory';
 
-import ViewMore from '@/components/Approval/components/Actions/components/ViewMore';
-import { AssetAvatar } from '@/components/AssetAvatar';
-import { toast } from '@/components2024/Toast';
+import { Text } from '@/components/Typography';
 import { RootNames } from '@/constant/layout';
+import { Account } from '@/core/services/preference';
 import { useAccounts } from '@/hooks/account';
 import { useSortAddressList } from '@/screens/Address/useSortAddressList';
-import { TransactionPendingDetail } from '@/screens/TransactionRecord/components/TransactionPendingDetail';
-import { ellipsisAddress } from '@/utils/address';
 import { naviPush } from '@/utils/navigation';
-import { formatIntlTimestamp } from '@/utils/time';
-import { openTxExternalUrl } from '@/utils/transaction';
 import { RevokeTokenApproveRequireData } from '@rabby-wallet/rabby-action';
 import { useMemoizedFn } from 'ahooks';
 import { unionBy } from 'lodash';
 import { useTranslation } from 'react-i18next';
-import { AddressItemInDetail } from '../../HistoryDetailScreen';
-import { TxStatusItem } from '../TxStatusItem';
 import { HistoryItemIcon } from '../HistoryItemIcon';
 import { HistoryItemCateType } from '../type';
-import { Account } from '@/core/services/preference';
-import { Text } from '@/components/Typography';
+import {
+  ActionDetailItem,
+  ActionDetailSection,
+  ActionDetailText,
+} from './components/ActionDetailSection';
+import { ActionSpenderView } from './components/ActionSpenderView';
 
 interface Props {
   data: TransactionGroup;
@@ -75,16 +69,6 @@ export const RevokeToken: React.FC<Props> = ({
     return unionBy(list, account => account.address.toLowerCase());
   }, [list]);
 
-  const handleOpenTxId = useMemoizedFn(() => {
-    const tx = data.maxGasTx.hash;
-
-    if (chain?.scanLink) {
-      openTxExternalUrl({ chain, txHash: tx });
-    } else {
-      toast.error('Unknown chain');
-    }
-  });
-
   const handleGotoTokenDetail = useMemoizedFn(() => {
     naviPush(RootNames.TokenDetail, {
       token: {
@@ -100,250 +84,67 @@ export const RevokeToken: React.FC<Props> = ({
   return (
     <>
       <TouchableOpacity onPress={handleGotoTokenDetail}>
-        <View style={[styles.singleBox]}>
-          <View style={{ flexDirection: 'row', alignItems: 'center', gap: 8 }}>
-            <HistoryItemIcon
-              isInDetail={true}
-              type={HistoryItemCateType.Revoke}
-              token={actionData.token}
-              isNft={false}
-            />
-            <View style={[styles.colomnBox]}>
-              <Text style={[styles.tokenAmountText, styles.isSendTextColor]}>
-                {getTokenSymbol(actionData.token as TokenItem)}
-              </Text>
+        <View style={styles.card}>
+          <View style={[styles.singleBox]}>
+            <View
+              style={{ flexDirection: 'row', alignItems: 'center', gap: 8 }}>
+              <HistoryItemIcon
+                isInDetail={true}
+                type={HistoryItemCateType.Revoke}
+                token={actionData.token}
+                isNft={false}
+              />
+              <View style={[styles.colomnBox]}>
+                <Text style={[styles.tokenAmountText, styles.isSendTextColor]}>
+                  {getTokenSymbol(actionData.token as TokenItem)}
+                </Text>
+              </View>
+            </View>
+            <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+              <RcIconSingleArrow
+                width={26}
+                height={26}
+                color={colors2024['neutral-bg-2']}
+              />
             </View>
           </View>
-          <View style={{ flexDirection: 'row', alignItems: 'center' }}>
-            <RcIconSingleArrow
-              width={26}
-              height={26}
-              color={colors2024['neutral-bg-2']}
+          <View style={styles.extraItem}>
+            <Text style={styles.itemTitleText}>
+              {t('page.transactions.detail.RevokeFrom')}
+            </Text>
+            <ActionSpenderView
+              requireData={requireData}
+              spender={actionData.spender}
+              chain={chain}
             />
           </View>
         </View>
       </TouchableOpacity>
-      <View style={styles.detailContainer}>
-        <View style={styles.detailContainerHeader}>
-          <Text style={styles.detailContainerTitle}>
-            {t('page.transactions.detail.TransactionDetails')}
-          </Text>
-        </View>
-        {!data.isPending && data.maxGasTx.completedAt && (
-          <View style={styles.detailItem}>
-            <Text style={styles.itemTitleText}>
-              {t('page.transactions.detail.Date')}
-            </Text>
-            <View>
-              <Text style={styles.itemContentText}>
-                {formatIntlTimestamp(data?.maxGasTx.completedAt)}
-              </Text>
-            </View>
-          </View>
-        )}
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.Status')}
-          </Text>
-          <View>
-            <TxStatusItem
-              status={data.isFailed ? 0 : 1}
-              isPending={data.isPending}
-              withText={true}
-            />
-          </View>
-        </View>
-        {data.isPending ? <TransactionPendingDetail data={data} /> : null}
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.RevokeFrom')}
-          </Text>
-          <ViewMore
-            type="spender"
-            data={{
-              ...requireData,
-              spender: actionData.spender,
-              chain,
-            }}>
-            <View style={{ alignItems: 'flex-end' }}>
-              <View
-                style={{
-                  flexDirection: 'row',
-                  alignItems: 'center',
-                  gap: 4,
-                }}>
-                <AssetAvatar logo={requireData?.protocol?.logo_url} size={16} />
-                <Text style={[styles.itemContentText]}>
-                  {requireData?.protocol?.name || t('global.Unknown')}
-                </Text>
-                <RcIconRightCC
-                  width={14}
-                  height={14}
-                  color={colors2024['neutral-foot']}
-                />
-              </View>
-              <Text style={styles.itemAddressText}>
-                {ellipsisAddress(actionData.spender)}
-              </Text>
-            </View>
-          </ViewMore>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.RevokeToken')}
-          </Text>
-          <Text style={styles.itemContentText}>
-            {getTokenSymbol(actionData.token)}
-          </Text>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.Chain')}
-          </Text>
-          <View style={{ flexDirection: 'row', gap: 4 }}>
-            <ChainIconImage
-              size={16}
-              chainEnum={chain?.enum}
-              isShowRPCStatus={true}
-            />
-            <Text style={[styles.itemContentText]}>{chain?.name}</Text>
-          </View>
-        </View>
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>
-            {t('page.transactions.detail.From')}
-          </Text>
-          <AddressItemInDetail
-            address={data.maxGasTx.address}
-            accounts={unionAccounts}
+      <ActionDetailSection data={data} chain={chain} accounts={unionAccounts}>
+        <ActionDetailItem label={t('page.transactions.detail.RevokeFrom')}>
+          <ActionSpenderView
+            requireData={requireData}
+            spender={actionData.spender}
+            chain={chain}
           />
-        </View>
+        </ActionDetailItem>
 
-        {Boolean(data.maxGasTx?.gasUSDValue) && (
-          <View style={styles.detailItem}>
-            <Text style={styles.itemTitleText}>
-              {t('page.transactions.detail.GasFee')}
-            </Text>
-            <Text style={styles.itemContentText}>
-              {formatAmount(data.maxGasTx?.gasTokenCount!)}{' '}
-              {data.maxGasTx?.gasTokenSymbol || ''} ($
-              {formatAmount(data.maxGasTx?.gasUSDValue ?? 0)})
-            </Text>
-          </View>
-        )}
-
-        <View style={styles.detailItem}>
-          <Text style={styles.itemTitleText}>Hash</Text>
-          <TouchableOpacity
-            disabled={!chain?.scanLink}
-            onPress={handleOpenTxId}
-            style={{ flexDirection: 'row', alignItems: 'center', gap: 2 }}>
-            <Text style={[styles.itemContentText]}>
-              {ellipsisAddress(data.maxGasTx.hash!)}
-            </Text>
-            <RcIconExternalLinkCC
-              width={14}
-              height={14}
-              color={colors2024['neutral-foot']}
-            />
-          </TouchableOpacity>
-        </View>
-      </View>
+        <ActionDetailItem label={t('page.transactions.detail.RevokeToken')}>
+          <ActionDetailText>
+            {getTokenSymbol(actionData.token)}
+          </ActionDetailText>
+        </ActionDetailItem>
+      </ActionDetailSection>
     </>
   );
 };
 
 const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
-  detailContainer: {
-    // flex: 1,
-    width: '100%',
-    marginTop: 12,
-    borderRadius: 16,
-    paddingTop: 12,
-    paddingBottom: 4,
-    backgroundColor: !isLight
-      ? colors2024['neutral-bg-2']
-      : colors2024['neutral-bg-1'],
-  },
-  detailContainerHeader: {
-    marginBottom: 8,
-    paddingHorizontal: 16,
-  },
-  detailContainerTitle: {
-    color: colors2024['neutral-body'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-  },
-  ghostButton: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderColor: colors2024['neutral-info'],
-  },
-  primaryButton: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderColor: colors2024['brand-default'],
-  },
-  primaryTitle: {
-    color: colors2024['brand-default'],
-  },
-  ghostTitle: {
-    color: colors2024['neutral-title-1'],
-  },
-  iconSwitchArrow: {
-    backgroundColor: colors2024['neutral-bg-2'],
-    borderRadius: 200,
-    width: 45,
-    height: 45,
-    alignItems: 'center',
-    justifyContent: 'center',
-    position: 'absolute',
-    left: '50%',
-    top: '50%',
-    marginLeft: -22,
-    marginTop: -22,
-  },
-  tokenAmountTextList: {
-    color: colors2024['green-default'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 18,
-    lineHeight: 22,
-    fontWeight: '700',
-  },
   colomnBox: {
     flexDirection: 'column',
   },
   isSendTextColor: {
     color: colors2024['neutral-title-1'],
-  },
-  isFailBox: {
-    opacity: 0.3,
-  },
-  image: {
-    width: 46,
-    height: 46,
-  },
-  fromTokenBox: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 16,
-    backgroundColor: colors2024['neutral-bg-1'],
-    flex: 1,
-    height: 110,
-    gap: 10,
-  },
-  toTokenBox: {
-    gap: 10,
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderRadius: 16,
-    backgroundColor: colors2024['neutral-bg-1'],
-    flex: 1,
-    height: 110,
   },
   card: {
     width: '100%',
@@ -373,40 +174,7 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     fontWeight: '500',
     marginTop: 2,
   },
-  mutliBox: {
-    width: '100%',
-    backgroundColor: colors2024['neutral-bg-1'],
-    justifyContent: 'center',
-    alignContent: 'center',
-    borderRadius: 16,
-    paddingHorizontal: 12,
-    paddingVertical: 16,
-    // flexDirection: 'row',
-    gap: 12,
-  },
-  doubleBox: {
-    justifyContent: 'center',
-    alignContent: 'center',
-    flexDirection: 'row',
-    height: 110,
-    gap: 10,
-    position: 'relative',
-  },
 
-  buttonContainer: {
-    position: 'absolute',
-    flexDirection: 'row',
-    height: 60,
-    bottom: 40,
-    width: '100%',
-    gap: 16,
-    left: 16,
-  },
-  itemAliaName: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 4,
-  },
   extraItem: {
     flexDirection: 'row',
     padding: 12,
@@ -417,14 +185,6 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     marginHorizontal: 12,
     marginBottom: 12,
   },
-  detailItem: {
-    flexDirection: 'row',
-    gap: 8,
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
   itemTitleText: {
     color: colors2024['neutral-secondary'],
     fontFamily: 'SF Pro Rounded',
@@ -433,36 +193,4 @@ const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
     fontWeight: '500',
     maxWidth: '45%',
   },
-  itemAddressText: {
-    color: colors2024['neutral-secondary'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '400',
-  },
-  itemContentText: {
-    color: colors2024['neutral-body'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-  },
-  headerTitleStyle: {
-    color: colors2024['neutral-title-1'],
-    fontWeight: '800',
-    fontSize: 20,
-    fontFamily: 'SF Pro Rounded',
-    lineHeight: 24,
-  },
-
-  statuItemText: {
-    color: colors2024['green-default'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-    marginLeft: 4,
-  },
-
-  headerItem: {},
 }));
