+++ apps/mobile/src/screens/Transaction/components/Actions/components/RevokeNFTBtn.tsx
@@ -1,20 +1,11 @@
-import { Tip } from '@/components';
-import { useSafeSetNavigationOptions } from '@/components/AppStatusBar';
-import { Button } from '@/components2024/Button';
 import { getErc721Approved, revokeNFTApprove } from '@/core/apis/approvals';
 import { KeyringAccountWithAlias } from '@/hooks/account';
-import { resetNavigationTo } from '@/hooks/navigation';
-import { useTheme2024 } from '@/hooks/theme';
-import { useMiniSigner } from '@/hooks/useSigner';
-import { isAccountSupportMiniApproval } from '@/utils/account';
-import { createGetStyles2024 } from '@/utils/styles';
 import { isSameAddress } from '@rabby-wallet/base-utils/src/isomorphic/address';
-import { NFTItem, Tx } from '@rabby-wallet/rabby-api/dist/types';
-import { useMemoizedFn, useRequest } from 'ahooks';
-import React, { useRef } from 'react';
-import { useTranslation } from 'react-i18next';
-import { View } from 'react-native';
-import { Text } from '@/components/Typography';
+import { NFTItem } from '@rabby-wallet/rabby-api/dist/types';
+import { useRequest } from 'ahooks';
+import React from 'react';
+import { RevokeApprovalCard } from './RevokeApprovalCard';
+import { useRevokeApproval } from './useRevokeApproval';
 
 interface Props {
   nft: NFTItem;
@@ -23,17 +14,6 @@ interface Props {
 }
 
 export const RevokeNFTBtn = ({ nft, spender, account }: Props) => {
-  const { t } = useTranslation();
-  const { styles, colors2024 } = useTheme2024({ getStyle });
-  const {
-    openUI,
-    resetGasStore,
-    close: closeMiniSign,
-  } = useMiniSigner({
-    account,
-  });
-  const { navigation } = useSafeSetNavigationOptions();
-
   const { data: isApproved } = useRequest(async () => {
     const approvedToAddress = await getErc721Approved({
       chainServerId: nft.chain,
@@ -45,9 +25,10 @@ export const RevokeNFTBtn = ({ nft, spender, account }: Props) => {
     return isSameAddress(spender, approvedToAddress);
   });
 
-  const handleRevokeDirectSign = useMemoizedFn(async () => {
-    try {
-      const data = await revokeNFTApprove(
+  const handleRevoke = useRevokeApproval({
+    account,
+    buildMiniSignTx: () =>
+      revokeNFTApprove(
         {
           chainServerId: nft.chain,
           nftTokenId: nft.inner_id,
@@ -59,33 +40,9 @@ export const RevokeNFTBtn = ({ nft, spender, account }: Props) => {
         },
         undefined,
         true,
-      );
-      const tx = data.params[0] as Tx;
-      closeMiniSign();
-      resetGasStore();
-      const res = await openUI({
-        txs: [tx],
-      });
-    } catch (e) {
-      console.error(e);
-      return;
-    }
-
-    if (navigation.canGoBack()) {
-      navigation.goBack();
-    } else {
-      resetNavigationTo(navigation, 'Home');
-    }
-  });
-
-  const handleRevoke = useMemoizedFn(async () => {
-    try {
-      if (isAccountSupportMiniApproval(account.type)) {
-        await handleRevokeDirectSign();
-        return;
-      }
-
-      await revokeNFTApprove({
+      ),
+    revoke: () =>
+      revokeNFTApprove({
         chainServerId: nft.chain,
         nftTokenId: nft.inner_id,
         spender: spender!,
@@ -93,104 +50,14 @@ export const RevokeNFTBtn = ({ nft, spender, account }: Props) => {
         abi: 'ERC721',
         isApprovedForAll: false,
         account: account,
-      });
-    } catch (e) {
-      console.error(e);
-    }
-
-    resetNavigationTo(navigation, 'Home');
+      }),
   });
 
-  const timerRef = useRef<ReturnType<typeof setTimeout>>(undefined);
-
   return (
-    <View style={styles.card}>
-      <View style={styles.cardHeader}>
-        <Text style={styles.label}>
-          {t('page.transactions.detail.totalApprovedAmount')}
-        </Text>
-        <Text style={styles.value}>{isApproved ? 1 : 0}</Text>
-      </View>
-      <View style={styles.buttonContainer}>
-        <Tip
-          placement="top"
-          pressableProps={{
-            onPress(ctx) {
-              ctx.turnOn();
-              if (timerRef.current) {
-                clearTimeout(timerRef.current);
-              }
-              timerRef.current = setTimeout(() => {
-                ctx.turnOff();
-              }, 3000);
-            },
-          }}
-          content={
-            !isApproved
-              ? t('page.transactions.detail.NoApproveNeed')
-              : undefined
-          }>
-          <Button
-            // loading={btnLoading}
-            disabled={!isApproved}
-            // buttonStyle={[styles.ghostButton]}
-            // titleStyle={[
-            //   styles.ghostTitle,
-            //   !isApproved && styles.ghostDisableButton,
-            // ]}
-            onPress={handleRevoke}
-            type={'primary'}
-            title={`${t('page.transactions.detail.Revoke')}`}
-          />
-        </Tip>
-      </View>
-    </View>
+    <RevokeApprovalCard
+      disabled={!isApproved}
+      onPress={handleRevoke}
+      value={isApproved ? 1 : 0}
+    />
   );
 };
-
-const getStyle = createGetStyles2024(({ colors2024, isLight }) => ({
-  card: {
-    padding: 16,
-    borderRadius: 16,
-    backgroundColor: !isLight
-      ? colors2024['neutral-bg-2']
-      : colors2024['neutral-bg-1'],
-    marginTop: 12,
-  },
-  cardHeader: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 12,
-    marginBottom: 16,
-  },
-
-  label: {
-    color: colors2024['neutral-secondary'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '500',
-    marginRight: 'auto',
-  },
-
-  value: {
-    color: colors2024['neutral-body'],
-    fontFamily: 'SF Pro Rounded',
-    fontSize: 14,
-    lineHeight: 18,
-    fontWeight: '700',
-  },
-
-  ghostButton: {
-    backgroundColor: colors2024['brand-light-1'],
-    borderColor: colors2024['brand-disable'],
-  },
-  ghostDisableButton: {
-    color: colors2024['brand-disable'],
-  },
-  ghostTitle: {
-    color: colors2024['brand-default'],
-  },
-  buttonContainer: {},
-}));
